+++ src/layouts/MainScreen/emotion.ts
@@ -63,14 +63,14 @@ export const DrawerHeader = styled("div")(({ theme }) => ({
 }));
 
 export const Profile = styled("div")`
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-  color: #000;
+	display: flex;
+	justify-content: flex-end;
+	width: 100%;
+	color: #000;
 `;
 
 export const MainBox = styled(Box)`
-  display: flex;
-  max-width: 100vw;
+	display: flex;
+	max-width: 100vw;
 	overflow-x: hidden;
 `;
