+++ src/pages/PlayerScreen/services.ts
@@ -1,6 +1,6 @@
 import { getTimer, newTimer, setTimer } from "api/timer";
 import { AxiosResponse } from "axios";
-import { IRunningTime, ITimer } from "types/Timer";
+import { ITimer } from "types/Timer";
 import { IYouTubeVideo } from "types/YouTube";
 import { getVideoById } from "api/youtube";
 
@@ -24,17 +24,6 @@ export async function setTimeAction(time: number) {
 		console.error(e);
 	}
 }
-
-export function convertHoursToMilliseconds(hours: number) {
-	return hours * 60 * 60 * 1000;
-}
-
-//convert hours ,minutes and seconds with decimal hour
-export function convertIntoHours(time: IRunningTime) {
-	const { hours, minutes, seconds } = time;
-	return hours + minutes / 60 + seconds / 3600;
-}
-
 
 export async function getVideoByIdAction(id: string) {
 	const videoTemp: IYouTubeVideo = (await getVideoById(id as string)).data;
